Rename services to managers
