+++ app/Models/Team.php
@@ -2,8 +2,8 @@
 
 namespace App\Models;
 
+use App\Managers\TeamStatsManager;
 use App\Models\GameMatch;
-use App\Services\TeamStatsService;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\DB;
@@ -22,7 +22,7 @@ class Team extends Model
 
     protected $table = 'teams';
 
-    protected $teamStatsService = null;
+    protected $teamStatsManager = null;
     
     /*
     |--------------------------------------------------------------------------
@@ -34,17 +34,17 @@ class Team extends Model
     {
         parent::__construct($attributes);
 
-        $this->teamStatsService = new TeamStatsService;
+        $this->teamStatsManager = new TeamStatsManager;
     }
 
     public function getStats(): array
     {
-        return $this->teamStatsService->getStats($this);
+        return $this->teamStatsManager->getStats($this);
     }
 
     public function getStatsWithPrediction(): array
     {
-        return $this->teamStatsService->getStats($this, true, true);
+        return $this->teamStatsManager->getStats($this, true, true);
     }
 
     public static function regenerateTeams(): int
